codecs: msm8916-wcd: Fix digital volume
With commit 5babb012c847 ("ASoC: codecs: msm8916-wcd-digital: move
gains from SX_TLV to S8_TLV") in the Linux kernel the scaling of the
digital volume has changed somewhat. Now we need to set the volume to
84 to get the same effect of 0 dB gain. Otherwise it clamps the current
value of 128 to 124 which represents the the maximum volume +40 dB and
is way too loud.